Migrates to vanniktech maven publish plugin
Migrates from the deprecated `maven-publish` and `signing` plugins to the `com.vanniktech.maven.publish` plugin for publishing to Maven Central.

This change simplifies the configuration and improves the reliability of the publishing process by using a more modern and actively maintained plugin.
It also addresses issues related to GPG key handling in CI environments by directly providing the key and passphrase as environment variables, rather than importing the key.